+++ include/grpcpp/impl/codegen/rpc_service_method.h
@@ -31,8 +31,11 @@
 #include <grpcpp/impl/codegen/rpc_method.h>
 #include <grpcpp/impl/codegen/status.h>
 
-namespace grpc {
+namespace grpc_impl {
 class ServerContextBase;
+}  // namespace grpc_impl
+
+namespace grpc {
 namespace internal {
 /// Base class for running an RPC handler.
 class MethodHandler {
@@ -49,8 +52,8 @@ class MethodHandler {
     /// \param requester : used only by the callback API. It is a function
     ///        called by the RPC Controller to request another RPC (and also
     ///        to set up the state required to make that request possible)
-    HandlerParameter(Call* c, ::grpc::ServerContextBase* context, void* req,
-                     Status req_status, void* handler_data,
+    HandlerParameter(Call* c, ::grpc_impl::ServerContextBase* context,
+                     void* req, Status req_status, void* handler_data,
                      std::function<void()> requester)
         : call(c),
           server_context(context),
@@ -60,7 +63,7 @@ class MethodHandler {
           call_requester(std::move(requester)) {}
     ~HandlerParameter() {}
     Call* const call;
-    ::grpc::ServerContextBase* const server_context;
+    ::grpc_impl::ServerContextBase* const server_context;
     void* const request;
     const Status status;
     void* const internal_data;
