+++ include/grpcpp/impl/codegen/server_interceptor.h
@@ -26,8 +26,12 @@
 #include <grpcpp/impl/codegen/rpc_method.h>
 #include <grpcpp/impl/codegen/string_ref.h>
 
-namespace grpc {
+namespace grpc_impl {
 class ServerContextBase;
+}  // namespace grpc_impl
+
+namespace grpc {
+
 namespace internal {
 class InterceptorBatchMethodsImpl;
 }
@@ -76,7 +80,7 @@ class ServerRpcInfo {
 
   /// Return a pointer to the underlying ServerContext structure associated
   /// with the RPC to support features that apply to it
-  ServerContextBase* server_context() { return ctx_; }
+  grpc_impl::ServerContextBase* server_context() { return ctx_; }
 
  private:
   static_assert(Type::UNARY ==
@@ -92,7 +96,7 @@ class ServerRpcInfo {
                     static_cast<Type>(internal::RpcMethod::BIDI_STREAMING),
                 "violated expectation about Type enum");
 
-  ServerRpcInfo(ServerContextBase* ctx, const char* method,
+  ServerRpcInfo(grpc_impl::ServerContextBase* ctx, const char* method,
                 internal::RpcMethod::RpcType type)
       : ctx_(ctx), method_(method), type_(static_cast<Type>(type)) {}
 
@@ -123,14 +127,14 @@ class ServerRpcInfo {
     }
   }
 
-  ServerContextBase* ctx_ = nullptr;
+  grpc_impl::ServerContextBase* ctx_ = nullptr;
   const char* method_ = nullptr;
   const Type type_;
   std::atomic<intptr_t> ref_{1};
   std::vector<std::unique_ptr<experimental::Interceptor>> interceptors_;
 
   friend class internal::InterceptorBatchMethodsImpl;
-  friend class grpc::ServerContextBase;
+  friend class grpc_impl::ServerContextBase;
 };
 
 }  // namespace experimental
