+++ include/grpcpp/impl/codegen/sync_stream_impl.h
@@ -23,7 +23,7 @@
 #include <grpcpp/impl/codegen/client_context.h>
 #include <grpcpp/impl/codegen/completion_queue.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
-#include <grpcpp/impl/codegen/server_context.h>
+#include <grpcpp/impl/codegen/server_context_impl.h>
 #include <grpcpp/impl/codegen/service_type.h>
 #include <grpcpp/impl/codegen/status.h>
 
@@ -612,12 +612,12 @@ class ServerReader final : public ServerReaderInterface<R> {
 
  private:
   ::grpc::internal::Call* const call_;
-  ::grpc::ServerContext* const ctx_;
+  ServerContext* const ctx_;
 
   template <class ServiceType, class RequestType, class ResponseType>
   friend class ::grpc_impl::internal::ClientStreamingHandler;
 
-  ServerReader(::grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerReader(::grpc::internal::Call* call, ::grpc_impl::ServerContext* ctx)
       : call_(call), ctx_(ctx) {}
 };
 
@@ -687,12 +687,12 @@ class ServerWriter final : public ServerWriterInterface<W> {
 
  private:
   ::grpc::internal::Call* const call_;
-  ::grpc::ServerContext* const ctx_;
+  ::grpc_impl::ServerContext* const ctx_;
 
   template <class ServiceType, class RequestType, class ResponseType>
   friend class ::grpc_impl::internal::ServerStreamingHandler;
 
-  ServerWriter(::grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerWriter(::grpc::internal::Call* call, ::grpc_impl::ServerContext* ctx)
       : call_(call), ctx_(ctx) {}
 };
 
@@ -707,7 +707,8 @@ namespace internal {
 template <class W, class R>
 class ServerReaderWriterBody final {
  public:
-  ServerReaderWriterBody(grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerReaderWriterBody(grpc::internal::Call* call,
+                         ::grpc_impl::ServerContext* ctx)
       : call_(call), ctx_(ctx) {}
 
   void SendInitialMetadata() {
@@ -766,7 +767,7 @@ class ServerReaderWriterBody final {
 
  private:
   grpc::internal::Call* const call_;
-  ::grpc::ServerContext* const ctx_;
+  ::grpc_impl::ServerContext* const ctx_;
 };
 
 }  // namespace internal
@@ -804,7 +805,8 @@ class ServerReaderWriter final : public ServerReaderWriterInterface<W, R> {
 
   friend class ::grpc_impl::internal::TemplatedBidiStreamingHandler<
       ServerReaderWriter<W, R>, false>;
-  ServerReaderWriter(::grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerReaderWriter(::grpc::internal::Call* call,
+                     ::grpc_impl::ServerContext* ctx)
       : body_(call, ctx) {}
 };
 
@@ -873,7 +875,8 @@ class ServerUnaryStreamer final
 
   friend class ::grpc_impl::internal::TemplatedBidiStreamingHandler<
       ServerUnaryStreamer<RequestType, ResponseType>, true>;
-  ServerUnaryStreamer(::grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerUnaryStreamer(::grpc::internal::Call* call,
+                      ::grpc_impl::ServerContext* ctx)
       : body_(call, ctx), read_done_(false), write_done_(false) {}
 };
 
@@ -934,7 +937,8 @@ class ServerSplitStreamer final
 
   friend class ::grpc_impl::internal::TemplatedBidiStreamingHandler<
       ServerSplitStreamer<RequestType, ResponseType>, false>;
-  ServerSplitStreamer(::grpc::internal::Call* call, ::grpc::ServerContext* ctx)
+  ServerSplitStreamer(::grpc::internal::Call* call,
+                      ::grpc_impl::ServerContext* ctx)
       : body_(call, ctx), read_done_(false) {}
 };
