+++ src/compiler/cpp_generator.cc
@@ -1345,7 +1345,7 @@ void PrintHeaderServerMethodStreamedUnary(
                    "  ::grpc::Service::MarkMethodStreamed($Idx$,\n"
                    "    new ::grpc::internal::StreamedUnaryHandler<\n"
                    "      $Request$, $Response$>(\n"
-                   "        [this](::grpc::ServerContext* context,\n"
+                   "        [this](::grpc_impl::ServerContext* context,\n"
                    "               ::grpc_impl::ServerUnaryStreamer<\n"
                    "                 $Request$, $Response$>* streamer) {\n"
                    "                   return this->Streamed$Method$(context,\n"
@@ -1399,7 +1399,7 @@ void PrintHeaderServerMethodSplitStreaming(
                    "  ::grpc::Service::MarkMethodStreamed($Idx$,\n"
                    "    new ::grpc::internal::SplitServerStreamingHandler<\n"
                    "      $Request$, $Response$>(\n"
-                   "        [this](::grpc::ServerContext* context,\n"
+                   "        [this](::grpc_impl::ServerContext* context,\n"
                    "               ::grpc_impl::ServerSplitStreamer<\n"
                    "                 $Request$, $Response$>* streamer) {\n"
                    "                   return this->Streamed$Method$(context,\n"
@@ -2257,7 +2257,7 @@ void PrintSourceService(grpc_generator::Printer* printer,
           "$Request$, "
           "$Response$>(\n"
           "        []($ns$$Service$::Service* service,\n"
-          "           ::grpc::ServerContext* ctx,\n"
+          "           ::grpc_impl::ServerContext* ctx,\n"
           "           const $Request$* req,\n"
           "           $Response$* resp) {\n"
           "             return service->$Method$(ctx, req, resp);\n"
@@ -2271,7 +2271,7 @@ void PrintSourceService(grpc_generator::Printer* printer,
           "    new ::grpc::internal::ClientStreamingHandler< "
           "$ns$$Service$::Service, $Request$, $Response$>(\n"
           "        []($ns$$Service$::Service* service,\n"
-          "           ::grpc::ServerContext* ctx,\n"
+          "           ::grpc_impl::ServerContext* ctx,\n"
           "           ::grpc_impl::ServerReader<$Request$>* reader,\n"
           "           $Response$* resp) {\n"
           "             return service->$Method$(ctx, reader, resp);\n"
@@ -2285,7 +2285,7 @@ void PrintSourceService(grpc_generator::Printer* printer,
           "    new ::grpc::internal::ServerStreamingHandler< "
           "$ns$$Service$::Service, $Request$, $Response$>(\n"
           "        []($ns$$Service$::Service* service,\n"
-          "           ::grpc::ServerContext* ctx,\n"
+          "           ::grpc_impl::ServerContext* ctx,\n"
           "           const $Request$* req,\n"
           "           ::grpc_impl::ServerWriter<$Response$>* writer) {\n"
           "             return service->$Method$(ctx, req, writer);\n"
@@ -2298,7 +2298,7 @@ void PrintSourceService(grpc_generator::Printer* printer,
                      "    new ::grpc::internal::BidiStreamingHandler< "
                      "$ns$$Service$::Service, $Request$, $Response$>(\n"
                      "        []($ns$$Service$::Service* service,\n"
-                     "           ::grpc::ServerContext* ctx,\n"
+                     "           ::grpc_impl::ServerContext* ctx,\n"
                      "           ::grpc_impl::ServerReaderWriter<$Response$,\n"
                      "           $Request$>* stream) {\n"
                      "             return service->$Method$(ctx, stream);\n"
