+++ src/cpp/server/server_context.cc
@@ -16,7 +16,7 @@
  *
  */
 
-#include <grpcpp/impl/codegen/server_context.h>
+#include <grpcpp/impl/codegen/server_context_impl.h>
 
 #include <algorithm>
 #include <utility>
@@ -35,16 +35,16 @@
 #include "src/core/lib/gprpp/sync.h"
 #include "src/core/lib/surface/call.h"
 
-namespace grpc {
+namespace grpc_impl {
 
 // CompletionOp
 
 class ServerContextBase::CompletionOp final
-    : public internal::CallOpSetInterface {
+    : public ::grpc::internal::CallOpSetInterface {
  public:
   // initial refs: one in the server context, one in the cq
   // must ref the call before calling constructor and after deleting this
-  CompletionOp(internal::Call* call,
+  CompletionOp(::grpc::internal::Call* call,
                ::grpc_impl::internal::ServerCallbackCall* callback_controller)
       : call_(*call),
         callback_controller_(callback_controller),
@@ -68,7 +68,7 @@ class ServerContextBase::CompletionOp final
     }
   }
 
-  void FillOps(internal::Call* call) override;
+  void FillOps(::grpc::internal::Call* call) override;
 
   // This should always be arena allocated in the call, so override delete.
   // But this class is not trivially destructible, so must actually call delete
@@ -89,7 +89,7 @@ class ServerContextBase::CompletionOp final
 
   bool FinalizeResult(void** tag, bool* status) override;
 
-  bool CheckCancelled(CompletionQueue* cq) {
+  bool CheckCancelled(grpc::CompletionQueue* cq) {
     cq->TryPluck(this);
     return CheckCancelledNoPluck();
   }
@@ -136,7 +136,7 @@ class ServerContextBase::CompletionOp final
     return finalized_ ? (cancelled_ != 0) : false;
   }
 
-  internal::Call call_;
+  ::grpc::internal::Call call_;
   ::grpc_impl::internal::ServerCallbackCall* const callback_controller_;
   bool has_tag_;
   void* tag_;
@@ -146,7 +146,7 @@ class ServerContextBase::CompletionOp final
   bool finalized_;
   int cancelled_;  // This is an int (not bool) because it is passed to core
   bool done_intercepting_;
-  internal::InterceptorBatchMethodsImpl interceptor_methods_;
+  ::grpc::internal::InterceptorBatchMethodsImpl interceptor_methods_;
 };
 
 void ServerContextBase::CompletionOp::Unref() {
@@ -157,7 +157,7 @@ void ServerContextBase::CompletionOp::Unref() {
   }
 }
 
-void ServerContextBase::CompletionOp::FillOps(internal::Call* call) {
+void ServerContextBase::CompletionOp::FillOps(::grpc::internal::Call* call) {
   grpc_op ops;
   ops.op = GRPC_OP_RECV_CLOSE_ON_SERVER;
   ops.data.recv_close_on_server.cancelled = &cancelled_;
@@ -206,7 +206,7 @@ bool ServerContextBase::CompletionOp::FinalizeResult(void** tag, bool* status) {
   }
   /* Add interception point and run through interceptors */
   interceptor_methods_.AddInterceptionHookPoint(
-      experimental::InterceptionHookPoints::POST_RECV_CLOSE);
+      ::grpc::experimental::InterceptionHookPoints::POST_RECV_CLOSE);
   if (interceptor_methods_.RunInterceptors()) {
     // No interceptors were run
     bool has_tag = has_tag_;
@@ -280,7 +280,7 @@ void ServerContextBase::Clear() {
 }
 
 void ServerContextBase::BeginCompletionOp(
-    internal::Call* call, std::function<void(bool)> callback,
+    ::grpc::internal::Call* call, std::function<void(bool)> callback,
     ::grpc_impl::internal::ServerCallbackCall* callback_controller) {
   GPR_ASSERT(!completion_op_);
   if (rpc_info_) {
@@ -301,8 +301,8 @@ void ServerContextBase::BeginCompletionOp(
   call->PerformOps(completion_op_);
 }
 
-internal::CompletionQueueTag* ServerContextBase::GetCompletionOpTag() {
-  return static_cast<internal::CompletionQueueTag*>(completion_op_);
+::grpc::internal::CompletionQueueTag* ServerContextBase::GetCompletionOpTag() {
+  return static_cast<::grpc::internal::CompletionQueueTag*>(completion_op_);
 }
 
 void ServerContextBase::AddInitialMetadata(const std::string& key,
@@ -316,7 +316,7 @@ void ServerContextBase::AddTrailingMetadata(const std::string& key,
 }
 
 void ServerContextBase::TryCancel() const {
-  internal::CancelInterceptorBatchMethods cancel_methods;
+  ::grpc::internal::CancelInterceptorBatchMethods cancel_methods;
   if (rpc_info_) {
     for (size_t i = 0; i < rpc_info_->interceptors_.size(); i++) {
       rpc_info_->RunInterceptor(&cancel_methods, i);
@@ -378,4 +378,4 @@ void ServerContextBase::SetLoadReportingCosts(
   }
 }
 
-}  // namespace grpc
+}  // namespace grpc_impl
